requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.28.30
botocore==1.31.30
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
defusedxml==0.7.1
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.101.1
fastjsonschema==2.18.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
langchain==0.0.268
langsmith==0.0.25
MarkupSafe==2.1.3
marshmallow==3.20.1
mistune==3.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandocfilters==1.5.0
platformdirs==3.10.0
pydantic==2.2.1
pydantic_core==2.6.1
Pygments==2.16.1
pypdf==3.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.9.2
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
webencodings==0.5.1
yarl==1.9.2
